pyproject.toml
[tool.poetry]
name = "gif"
version = "0.1.0"
description = "Making gifs from still images."
authors = ["Aatif Syed <aatif.syed@metaswitch.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Wand = "^0.6.5"
Pillow = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
ipython = "^7.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "visual: requires tester inspection"
]